fix(swiftlint): extract ActionRunsResponse DTOs to drop ActionGroup.swift under 400 lines
